Deleted random_value.py, which was left over from the Messaging test split. Removed unused functions from _common.sh.
Removed the getEntity function, as it is now redundant.

Added function documentation to the functions in _common.sh.

Added function documentation and made formatting improvements in SalGen.sh.

Added a comment heading to variables.py.

Archived three Jenkinsfiles, as they are not used anymore, but should be kept around for a bit longer, just in case.