Added the first tests of the SAL AuthList feature.
Added the version parameter to the Java salgenerator process.

Added the C++ AuthList tests to Jenkinsfile.quick.

Regenerated the test suites with the new AuthList checks.

Turns out the Script does not support the Java language, so there are no Java AuthList tests to run in the SAL/XML quick tests, Jenkinsfile.quick. Restructured the block to only run the C++ AuthList tests.

Added the AuthList tests to Jenkinsfile.salxml. They are run in a separate parallel stage block, after the messaging tests.

Too much cross-talk in the AuthList tests when running in parallel; updated Jenkinsfile.salxml to run them in series.